fix: run JMeter loading with property file gotten earlier when Nacos service is not available

+++ bin/debug.py
@@ -28,11 +28,21 @@ if __name__ == "__main__":
     os.chdir(snapshot_base)
     old_file = "+".join([stage, summary_group, settings.SUMMARY_NAMESPACE_ID])
     new_file = f"{stage}.properties"
-    # delete if already exist
-    if os.path.isfile(new_file):
-        os.remove(new_file)
-    os.rename(old_file, new_file)
 
-    # start JMeter with additional properties
-    logger.info(f"start JMeter with property file {snapshot_base}/{new_file} loaded")
-    subprocess.run(f"jmeter -q {new_file}", shell=True)
+    # check if get config file successfully
+    if os.path.isfile(old_file):
+        # delete if already exist
+        if os.path.isfile(new_file):
+            os.remove(new_file)
+        os.rename(old_file, new_file)
+        # start JMeter with additional properties
+        logger.info(f"start JMeter with property file {snapshot_base}/{new_file} loaded")
+        subprocess.run(f"jmeter -q {new_file}", shell=True)
+    else:
+        if os.path.isfile(new_file):
+            logger.warning("Nacos 服务好像挂了，无法拿到最新的配置，加载之前的配置文件来启动 JMeter")
+            # start JMeter with additional properties
+            logger.info(f"start JMeter with property file {snapshot_base}/{new_file} loaded")
+            subprocess.run(f"jmeter -q {new_file}", shell=True)
+        else:
+            logger.error("Nacos 服务好像挂了，无法拿到最新的配置")
